rename-tax-invoices/requirements.txt
PyMuPDF>=1.24.0
google-genai>=1.0.0
watchdog>=4.0.0
